sd_interface/requirements.txt
fastapi==0.110.1
Pillow==10.3.0
pydantic==2.7.0
Requests==2.31.0
uvicorn==0.29.0
websocket_client==1.7.0
